+++ lib/src/iterable/repeat.dart
@@ -4,39 +4,62 @@ import 'dart:collection' show IterableBase;
 
 import 'mixins/infinite.dart';
 
+extension RepeatExtension<E> on Iterable<E> {
+  /// Returns an infinite iterable with the elements of this iterable. If
+  /// [count] is provided the resulting iterator is limited to [count]
+  /// repetitions.
+  ///
+  /// Example expressions:
+  ///
+  ///    [1, 2].repeat();               // [1, 2, 1, 2, ...]
+  ///    [1, 2, 3].repeat(count: 2);    // [1, 2, 3, 1, 2, 3]
+  ///
+  Iterable<E> repeat({int count}) {
+    if (isEmpty || count == 0) {
+      return const Iterable.empty();
+    } else if (count == null) {
+      return RepeatIterable(this);
+    } else if (this is InfiniteIterable) {
+      return this;
+    }
+    RangeError.checkNotNegative(count, 'count');
+    return RepeatIterable(this).take(length * count);
+  }
+}
+
 /// Returns an infinite iterable with a constant [element]. If [count] is
 /// provided the resulting iterator is limited to [count] elements.
 ///
 /// Example expressions:
 ///
-///     repeat(2);        // [2, 2, 2, 2, 2, 2, ...]
-///     repeat('a', 3);   // ['a', 'a', 'a']
+///     repeat(2);               // [2, 2, 2, 2, 2, 2, ...]
+///     repeat('a', count: 3);   // ['a', 'a', 'a']
 ///
-Iterable<E> repeat<E>(E element, {int count}) {
-  final iterable = RepeatIterable<E>(element);
-  return count == null ? iterable : iterable.take(count);
-}
+Iterable<E> repeat<E>(E element, {int count}) => [element].repeat(count: count);
 
 class RepeatIterable<E> extends IterableBase<E> with InfiniteIterable<E> {
-  final E element;
+  final Iterable<E> iterable;
 
-  RepeatIterable(this.element);
+  RepeatIterable(this.iterable);
 
   @override
-  Iterator<E> get iterator => RepeatIterator<E>(element);
+  Iterator<E> get iterator => RepeatIterator<E>(iterable);
 }
 
 class RepeatIterator<E> extends Iterator<E> {
-  final E element;
+  final Iterable<E> iterable;
+  Iterator<E> iterator;
 
-  RepeatIterator(this.element);
+  RepeatIterator(this.iterable) : iterator = iterable.iterator;
 
   @override
-  E current;
+  E get current => iterator.current;
 
   @override
   bool moveNext() {
-    current = element;
+    while (!iterator.moveNext()) {
+      iterator = iterable.iterator;
+    }
     return true;
   }
 }