+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2.6.0
+
+* Add `repeat` function on `Iterable`.
+
 ## 2.5.0
 
 * Printer package correctly trims, pads and separates unicode strings now.
@@ -23,8 +27,8 @@
 * Made `truncateToPeriod` and `periodical` extension methods of `DateTime`.
 * Add extension methods to convert 
   * a `Map<K, V>` to a `BiMap<K, V>`, 
-  * an `Iterable<bool>` to a `BitList`, 
-  * and an `Iterable<T>` to a `Multiset<T>`.
+  * an `Iterable<bool>` to a `BitList`, and
+  * an `Iterable<T>` to a `Multiset<T>`.
 
 ## 2.1.0